Convert clickable image to input
Currently the way for somebody to open the main menu on this site is a click handler attached to an image. In order for a keyboard user to open the menu, that image should be a button. This PR changes the image to an input with type of "image" which will render a button that is keyboard-focusable and has the default visible browser focus style when it is focused. This means that somebody with a disability who needs to use the keyboard to get around the page, can now open the menu.