Remove non-descriptive alt text in profile cards
It's better to have an empty alt attribute in this case, so the image is not announced by screen readers, since the very next text to be read out is the speaker's name. Having the alt text by the word image will cause a screen reader to read out something like "Image of image" for every card, which would be tedious and add no information for the user.